+++ lib/chat_web/live/main_live/layout/admin.ex
@@ -76,4 +76,30 @@ defmodule ChatWeb.MainLive.Layout.Admin do
     </div>
     """
   end
+
+  attr :status, :boolean, required: true, doc: ":on/:off"
+
+  def gpio24_impedance_control(assigns) do
+    ~H"""
+    <div class="flex flex-col space-y-5">
+      <div class="mt-4">
+        <%= if @status do %>
+          <p class="text-gray-700">
+            Status: <span id="status-text" class="font-medium"><%= @status %></span>
+          </p>
+          <button
+            class="py-2 px-4 mt-3 w-full h-13 border-0 rounded-lg bg-grayscale text-white font-medium"
+            phx-click="admin/toggle-gpio24-impendance"
+          >
+            Toggle Impedance
+          </button>
+        <% else %>
+          <p class="text-gray-700">
+            Status: <span id="status-text" class="font-medium">Loading...</span>
+          </p>
+        <% end %>
+      </div>
+    </div>
+    """
+  end
 end